Session 3: binding fix, PSRAM allocation, type widening, name arrays
Binding fix (controls not propagating in FastLED-MM)
- resolveField replaces resolveCVar: binding targets are now emitter/flow
  struct fields directly instead of cVar globals
- Bindings applied after syncFromCVars() in run() so external values
  always win, even across emitter/flow switches that call pushDefaultsToCVars
- Fixes persistence double-counting (cPersistFine accumulation)

EMITTER/FLOW encapsulation
- Moved uint8_t EMITTER/FLOW globals into FlowFieldsEngine as _emitter/_flow
- bleControl.h reads/writes via g_engine->_emitter/_flow
- Library consumers no longer need to define these externally

PSRAM-aware grid allocation
- ps_malloc() used for all 8 float arrays on BOARD_HAS_PSRAM devices
- malloc() fallback for boards without PSRAM; free() valid for both
- Covers gR/gG/gB, tR/tG/tB, xProf, yProf

Dimension type widening
- _width/_height/_minDim: uint8_t -> uint16_t
- _numLeds: uint16_t -> uint32_t
- xyFunc signature: uint16_t(*)(uint8_t,uint8_t) -> uint32_t(*)(uint16_t,uint16_t)
- Supports panels up to 1024x1024 for PC targets

EMITTER_NAMES / FLOW_NAMES
- Added as inline const char* arrays to componentEnums.h
- Parallel to enum definitions; available without including BLE headers
- Used in FlowFieldsEffect.h to populate select option labels

Docs
- Session 3 retrospective added to FlowFieldsAsALibrary.md

+++ src/boardConfig.h
@@ -84,7 +84,7 @@ extern const uint16_t serpTopDown[NUM_LEDS] PROGMEM;
 extern const uint16_t serpBottomUp[NUM_LEDS] PROGMEM;
 
 extern uint8_t cMapping;
-extern uint16_t ledNum;
+extern uint32_t ledNum;
 
 enum Mapping {
 	TopDownProgressive = 0,
@@ -93,9 +93,9 @@ enum Mapping {
 	BottomUpSerpentine
 };
 
-uint16_t myXY(uint8_t x, uint8_t y) {
+uint32_t myXY(uint16_t x, uint16_t y) {
 		if (x >= WIDTH || y >= HEIGHT) return 0;
-		uint16_t i = ( y * WIDTH ) + x;
+		uint32_t i = (uint32_t)y * WIDTH + x;
 		switch(cMapping){
 			case 0:	 ledNum = progTopDown[i]; break;
 			case 1:	 ledNum = progBottomUp[i]; break;
